impl mapper with eager loaded entity, split query as multiple collections being returned

+++ src/DfE.EducationProviderRegistry.Core.Query/Groups/Application/Model/GroupUID.cs
@@ -2,13 +2,13 @@
 
 public readonly record struct GroupUID
 {
-    public GroupUID(int groupId)
+    public GroupUID(long groupId)
     {
         ArgumentOutOfRangeException.ThrowIfNegativeOrZero(groupId);
         Value = groupId;
     }
 
-    public int Value { get; }
+    public long Value { get; }
 
     public static bool TryCreate(string input, out GroupUID result)
     {
